requirements.txt
python-telegram-bot
pytz
openpyxl
unidecode
flask
flask_session
pygithub
gunicorn
telethon
pillow